improved retry handling in directory clear extension plus better error trapping readonly applied where possible

+++ Fastnet.Core/FileSystem/ClearDirectory.cs
@@ -5,46 +5,58 @@ using System.Threading;
 
 namespace Fastnet.Core
 {
-    public static partial class extensions
+    public static partial class Extensions
     {
         /// <summary>
         /// 
         /// </summary>
         /// <param name="dir"></param>
         /// <param name="log"></param>
-        public static void Clear(this DirectoryInfo dir, ILogger log = null)
+        public static void Clear(this DirectoryInfo dir/*, ILogger log = null*/)
         {
-            foreach (var file in dir.EnumerateFiles())
+            if (dir.Exists)
             {
-                try
+                foreach (var file in dir.EnumerateFiles())
                 {
-                    file.Delete();
-                    log?.Trace($"{file.FullName} deleted");
-                }
-                catch (Exception)
-                {
-                    log?.Error($"failed to delete {file.FullName}");
-                    throw;
-                }
-            }
-            foreach (var d in dir.EnumerateDirectories())
-            {
-                int retryCount = 1;
-                while (retryCount > 0)
-                {
-                    d.Clear(log);
                     try
                     {
-                        d.Delete();
-                        log?.Trace($"{d.FullName} deleted");
+                        file.Delete();
+                        log.Trace($"{file.FullName} deleted");
+                    }
+                    catch(FileNotFoundException)
+                    {
+
                     }
                     catch (Exception)
                     {
-                        log?.Error($"failed to delete {d.FullName}");
-                        Thread.Sleep(2000);
-                        --retryCount;
+                        log.Error($"failed to delete {file.FullName}");
+                        throw;
                     }
                 }
+                foreach (var d in dir.EnumerateDirectories())
+                {
+                    int retryCount = 10;
+                    while (retryCount > 0)
+                    {
+                        try
+                        {
+                            d.Clear();
+                            d.Delete();
+                            log.Trace($"{d.FullName} deleted");
+                            break;
+                        }
+                        catch (Exception)
+                        {
+                            log.Error($"failed to delete {d.FullName}, retry count is {retryCount}");
+                            Thread.Sleep(50);
+                            --retryCount;
+                        }
+                    }
+                } 
+            }
+            else
+            {
+                log.Warning($"directory {dir.FullName} unexpectedly not found");
             }
         }
     }
